add separate utils tests
